@@ -5,7 +5,7 @@
     "keywords": [
         "image", "exif"
     ],
-    "homepage": "http://lsolesen.github.com/pel/",
+    "homepage": "http://pel.github.com/pel/",
     "authors": [
         {
             "name": "Lars Olesen",
@@ -22,15 +22,18 @@
     ],
     "license": "GPL-2.0",
     "require": {
-        "php": ">=5.0.0"
+        "php": ">=5.4.0"
     },
     "require-dev" : {
         "ext-gd": "*",
-        "phpunit/phpunit": "5.7.*",
-        "squizlabs/php_codesniffer": "3.0.0RC3",
+        "phpunit/phpunit": ">=4.8.36 <8",
+        "squizlabs/php_codesniffer": "^3.0.0",
         "satooshi/php-coveralls": "1.0.*"
     },
     "autoload": {
         "psr-4": {"lsolesen\\pel\\": "src/"}
+    },
+    "autoload-dev": {
+        "psr-4": { "Pel\\Test\\": "test/" }
     }
 }
