@@ -3,6 +3,10 @@
 namespace Drupal\devel\Controller;
 
 use Drupal\Core\Controller\ControllerBase;
+use Drupal\Core\Entity\EntityTypeBundleInfoInterface;
+use Drupal\Core\Field\FieldTypePluginManagerInterface;
+use Drupal\Core\Field\FormatterPluginManager;
+use Drupal\Core\Field\WidgetPluginManager;
 use Drupal\Core\Url;
 use Drupal\devel\DevelDumperManagerInterface;
 use Drupal\field\Entity\FieldConfig;
@@ -22,20 +26,66 @@ class DevelController extends ControllerBase {
   protected $dumper;
 
   /**
+   * The entity type bundle info service.
+   *
+   * @var \Drupal\Core\Entity\EntityTypeBundleInfoInterface
+   */
+  protected $entityTypeBundleInfo;
+
+  /**
+   * The field type plugin manager service.
+   *
+   * @var \Drupal\Core\Field\FieldTypePluginManagerInterface
+   */
+  protected $fieldTypeManager;
+
+  /**
+   * The field formatter plugin manager.
+   *
+   * @var \Drupal\Core\Field\FormatterPluginManager
+   */
+  protected $formatterPluginManager;
+
+  /**
+   * The field widget plugin manager.
+   *
+   * @var \Drupal\Core\Field\WidgetPluginManager
+   */
+  protected $widgetPluginManager;
+
+  /**
    * EntityDebugController constructor.
    *
    * @param \Drupal\devel\DevelDumperManagerInterface $dumper
    *   The dumper service.
+   * @param \Drupal\Core\Entity\EntityTypeBundleInfoInterface $entity_type_bundle_info
+   *   The entity type bundle info service.
+   * @param \Drupal\Core\Field\FieldTypePluginManagerInterface $field_type_manager
+   *   The field type manager service.
+   * @param \Drupal\Core\Field\FormatterPluginManager $formatter_plugin_manager
+   *   The field formatter plugin manager.
+   * @param \Drupal\Core\Field\WidgetPluginManager $widget_plugin_manager
+   *   The field widget plugin manager.
    */
-  public function __construct(DevelDumperManagerInterface $dumper) {
+  public function __construct(DevelDumperManagerInterface $dumper, EntityTypeBundleInfoInterface $entity_type_bundle_info, FieldTypePluginManagerInterface $field_type_manager, FormatterPluginManager $formatter_plugin_manager, WidgetPluginManager $widget_plugin_manager) {
     $this->dumper = $dumper;
+    $this->entityTypeBundleInfo = $entity_type_bundle_info;
+    $this->fieldTypeManager = $field_type_manager;
+    $this->formatterPluginManager = $formatter_plugin_manager;
+    $this->widgetPluginManager = $widget_plugin_manager;
   }
 
   /**
    * {@inheritdoc}
    */
   public static function create(ContainerInterface $container) {
-    return new static($container->get('devel.dumper'));
+    return new static(
+      $container->get('devel.dumper'),
+      $container->get('entity_type.bundle.info'),
+      $container->get('plugin.manager.field.field_type'),
+      $container->get('plugin.manager.field.formatter'),
+      $container->get('plugin.manager.field.widget')
+    );
   }
 
   /**
@@ -47,6 +97,12 @@ class DevelController extends ControllerBase {
     return $this->redirect('<front>');
   }
 
+  /**
+   * Theme registry.
+   *
+   * @return array
+   *   The complete theme registry as renderable.
+   */
   public function themeRegistry() {
     $hooks = theme_get_registry();
     ksort($hooks);
@@ -68,19 +124,19 @@ class DevelController extends ControllerBase {
     ksort($field_instances);
     $output['instances'] = $this->dumper->exportAsRenderable($field_instances, $this->t('Instances'));
 
-    $bundles = \Drupal::service('entity_type.bundle.info')->getAllBundleInfo();
+    $bundles = $this->entityTypeBundleInfo->getAllBundleInfo();
     ksort($bundles);
     $output['bundles'] = $this->dumper->exportAsRenderable($bundles, $this->t('Bundles'));
 
-    $field_types = \Drupal::service('plugin.manager.field.field_type')->getUiDefinitions();
+    $field_types = $this->fieldTypeManager->getUiDefinitions();
     ksort($field_types);
     $output['field_types'] = $this->dumper->exportAsRenderable($field_types, $this->t('Field types'));
 
-    $formatter_types = \Drupal::service('plugin.manager.field.formatter')->getDefinitions();
+    $formatter_types = $this->formatterPluginManager->getDefinitions();
     ksort($formatter_types);
     $output['formatter_types'] = $this->dumper->exportAsRenderable($formatter_types, $this->t('Formatter types'));
 
-    $widget_types = \Drupal::service('plugin.manager.field.widget')->getDefinitions();
+    $widget_types = $this->widgetPluginManager->getDefinitions();
     ksort($widget_types);
     $output['widget_types'] = $this->dumper->exportAsRenderable($widget_types, $this->t('Widget types'));
 
@@ -153,15 +209,15 @@ class DevelController extends ControllerBase {
    *   Array of page elements to render.
    */
   public function session() {
-    $output['description'] = array(
+    $output['description'] = [
       '#markup' => '<p>' . $this->t('Here are the contents of your $_SESSION variable.') . '</p>',
-    );
-    $output['session'] = array(
+    ];
+    $output['session'] = [
       '#type' => 'table',
-      '#header' => array($this->t('Session name'), $this->t('Session ID')),
-      '#rows' => array(array(session_name(), session_id())),
+      '#header' => [$this->t('Session name'), $this->t('Session ID')],
+      '#rows' => [[session_name(), session_id()]],
       '#empty' => $this->t('No session available.'),
-    );
+    ];
     $output['data'] = $this->dumper->exportAsRenderable($_SESSION);
 
     return $output;
