@@ -92,8 +92,8 @@ class SettingsForm extends ConfigFormBase {
       '#options' => [
         DEVEL_ERROR_HANDLER_NONE => t('None'),
         DEVEL_ERROR_HANDLER_STANDARD => t('Standard Drupal'),
-        DEVEL_ERROR_HANDLER_BACKTRACE_DPM => t('Kint backtrace in the message area'),
-        DEVEL_ERROR_HANDLER_BACKTRACE_KINT => t('Kint backtrace above the rendered page'),
+        DEVEL_ERROR_HANDLER_BACKTRACE_DPM => t('Backtrace in the message area'),
+        DEVEL_ERROR_HANDLER_BACKTRACE_KINT => t('Backtrace above the rendered page'),
       ],
       '#multiple' => TRUE,
       '#default_value' => empty($error_handlers) ? DEVEL_ERROR_HANDLER_NONE : $error_handlers,
@@ -106,11 +106,11 @@ class SettingsForm extends ConfigFormBase {
           '#items' => [
             $this->t('<em>None</em> is a good option when stepping through the site in your debugger.'),
             $this->t('<em>Standard Drupal</em> does not display all the information that is often needed to resolve an issue.'),
-            $this->t('<em>Kint backtrace</em> displays nice debug information when any type of error is noticed, but only to users with the %perm permission.', ['%perm' => t('Access developer information')]),
+            $this->t('<em>Backtrace</em> displays nice debug information when any type of error is noticed, but only to users with the %perm permission.', ['%perm' => t('Access developer information')]),
           ],
         ],
         [
-          '#markup' => $this->t('Depending on the situation, the theme, the size of the call stack and the arguments, etc., some handlers may not display their messages, or display them on the subsequent page. Select <em>Standard Drupal</em> <strong>and</strong> <em>Kint backtrace above the rendered page</em> to maximize your chances of not missing any messages.') . '<br />' .
+          '#markup' => $this->t('Depending on the situation, the theme, the size of the call stack and the arguments, etc., some handlers may not display their messages, or display them on the subsequent page. Select <em>Standard Drupal</em> <strong>and</strong> <em>Backtrace above the rendered page</em> to maximize your chances of not missing any messages.') . '<br />' .
             $this->t('Demonstrate the current error handler(s):') . ' ' .
             $this->l('notice', $current_url->setOption('query', ['demo' => 'notice'])) . ', ' .
             $this->l('notice+warning', $current_url->setOption('query', ['demo' => 'warning'])) . ', ' .
@@ -136,9 +136,8 @@ class SettingsForm extends ConfigFormBase {
       '#title' => $this->t('Variables Dumper'),
       '#options' => [],
       '#default_value' => $default,
-      '#description' => $this->t('Select the debugging tool used for formatting and displaying the variables inspected through the debug functions of Devel. You can enable the <a href=":kint_install">Kint module</a> (shipped with Devel) and select the Kint debugging tool for an improved debugging experience. <strong>NOTE</strong>: Some of these plugins require external libraries for to be enabled. Learn how install external libraries with <a href=":url">Composer</a>.', [
+      '#description' => $this->t('Select the debugging tool used for formatting and displaying the variables inspected through the debug functions of Devel. <strong>NOTE</strong>: Some of these plugins require external libraries for to be enabled. Learn how install external libraries with <a href=":url">Composer</a>.', [
         ':url' => 'https://www.drupal.org/node/2404989',
-        ':kint_install' => Url::fromRoute('system.modules_list')->toString(),
       ]),
     ];
 
