@@ -103,8 +103,11 @@ class Debug extends \Twig_Extension {
       $this->dumper->dump($context_variables, 'Twig context');
     }
     else {
-      foreach ($args as $variable) {
-        $this->dumper->dump($variable);
+      $parameters = $this->guessTwigFunctionParameters();
+
+      foreach ($args as $index => $variable) {
+        $name = !empty($parameters[$index]) ? $parameters[$index] : NULL;
+        $this->dumper->dump($variable, $name);
       }
     }
 
@@ -136,8 +139,11 @@ class Debug extends \Twig_Extension {
       $this->dumper->message($context_variables, 'Twig context');
     }
     else {
-      foreach ($args as $variable) {
-        $this->dumper->message($variable);
+      $parameters = $this->guessTwigFunctionParameters();
+
+      foreach ($args as $index => $variable) {
+        $name = !empty($parameters[$index]) ? $parameters[$index] : NULL;
+        $this->dumper->message($variable, $name);
       }
     }
 
@@ -194,6 +200,52 @@ class Debug extends \Twig_Extension {
       }
     }
     return $context_variables;
+  }
+
+  /**
+   * Gets the twig function parameters for the current invocation.
+   *
+   * @return array
+   *   The detected twig function parameters.
+   */
+  protected function guessTwigFunctionParameters() {
+    $callee = NULL;
+    $template = NULL;
+
+    $backtrace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS | DEBUG_BACKTRACE_PROVIDE_OBJECT);
+
+    foreach ($backtrace as $index => $trace) {
+      if (isset($trace['object']) && $trace['object'] instanceof \Twig_Template && 'Twig_Template' !== get_class($trace['object'])) {
+        $template = $trace['object'];
+        $callee = $backtrace[$index - 1];
+        break;
+      }
+    }
+
+    $parameters = [];
+
+    /** @var \Twig_Template $template */
+    if (NULL !== $template && NULL !== $callee) {
+      $line_number = $callee['line'];
+      $debug_infos = $template->getDebugInfo();
+
+      if (isset($debug_infos[$line_number])) {
+        $source_line = $debug_infos[$line_number];
+        $source_file_name = $template->getTemplateName();
+
+        if (is_readable($source_file_name)) {
+          $source = file($source_file_name, FILE_IGNORE_NEW_LINES);
+          $line = $source[$source_line - 1];
+
+          preg_match('/\((.+)\)/', $line, $matches);
+          if (isset($matches[1])) {
+            $parameters = array_map('trim', explode(',', $matches[1]));
+          }
+        }
+      }
+    }
+
+    return $parameters;
   }
 
 }
