@@ -2,10 +2,12 @@
 
 namespace Drupal\Tests\file_mdm_exif\Kernel;
 
+use Drupal\Core\File\FileSystemInterface;
 use Drupal\file_mdm\FileMetadataInterface;
 use Drupal\Tests\file_mdm\Kernel\FileMetadataManagerTestBase;
 use lsolesen\pel\PelEntryAscii;
 use lsolesen\pel\PelEntryRational;
+use lsolesen\pel\PelEntryShort;
 use lsolesen\pel\PelEntrySRational;
 
 /**
@@ -22,7 +24,6 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
    */
   public static $modules = [
     'system',
-    'simpletest',
     'file_mdm',
     'file_mdm_exif',
     'file_test',
@@ -32,7 +33,7 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
   /**
    * {@inheritdoc}
    */
-  public function setUp() {
+  public function setUp(): void {
     parent::setUp();
     $this->installConfig(['file_mdm_exif']);
   }
@@ -42,16 +43,16 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
    */
   public function testExifPlugin() {
     // Prepare a copy of test files.
-    file_unmanaged_copy(drupal_get_path('module', 'simpletest') . '/files/image-test.jpg', 'public://', FILE_EXISTS_REPLACE);
-    file_unmanaged_copy(drupal_get_path('module', 'simpletest') . '/files/image-test.png', 'public://', FILE_EXISTS_REPLACE);
-    file_unmanaged_copy(drupal_get_path('module', 'file_mdm') . '/tests/files/test-exif.jpeg', 'public://', FILE_EXISTS_REPLACE);
-    file_unmanaged_copy(drupal_get_path('module', 'file_mdm') . '/tests/files/test-exif.jpeg', 'temporary://', FILE_EXISTS_REPLACE);
+    $this->fileSystem->copy('core/tests/fixtures/files/image-test.jpg', 'public://', FileSystemInterface::EXISTS_REPLACE);
+    $this->fileSystem->copy('core/tests/fixtures/files/image-test.png', 'public://', FileSystemInterface::EXISTS_REPLACE);
+    $this->fileSystem->copy(drupal_get_path('module', 'file_mdm') . '/tests/files/test-exif.jpeg', 'public://', FileSystemInterface::EXISTS_REPLACE);
+    $this->fileSystem->copy(drupal_get_path('module', 'file_mdm') . '/tests/files/test-exif.jpeg', 'temporary://', FileSystemInterface::EXISTS_REPLACE);
     // The image files that will be tested.
     $image_files = [
       [
         // Pass a path instead of the URI.
         'uri' => drupal_get_path('module', 'file_mdm') . '/tests/files/test-exif.jpeg',
-        'count_keys' => 48,
+        'count_keys' => 47,
         'test_keys' => [
           ['Orientation', 8],
           ['orientation', 8],
@@ -67,7 +68,7 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
       [
         // Pass a URI.
         'uri' => 'public://test-exif.jpeg',
-        'count_keys' => 48,
+        'count_keys' => 47,
         'test_keys' => [
           ['Orientation', 8],
           ['ShutterSpeedValue', [106, 32]],
@@ -77,7 +78,7 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
         // Remote storage file. Let the file be copied to a local temp copy.
         'uri' => 'dummy-remote://test-exif.jpeg',
         'copy_to_temp' => TRUE,
-        'count_keys' => 48,
+        'count_keys' => 47,
         'test_keys' => [
           ['Orientation', 8],
           ['ShutterSpeedValue', [106, 32]],
@@ -105,7 +106,7 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
       [
         // TIFF image.
         'uri' => drupal_get_path('module', 'file_mdm') . '/tests/files/sample-1.tiff',
-        'count_keys' => 11,
+        'count_keys' => 15,
         'test_keys' => [
           ['Orientation', 1],
           ['BitsPerSample', [8, 8, 8, 8]],
@@ -131,10 +132,10 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
       if (isset($image_file['copy_to_temp'])) {
         $file_metadata->copyUriToTemp();
       }
-      $this->assertEqual($image_file['count_keys'], $this->countMetadataKeys($file_metadata, 'exif'));
+      $this->assertEquals($image_file['count_keys'], $this->countMetadataKeys($file_metadata, 'exif'));
       foreach ($image_file['test_keys'] as $test) {
         $entry = $file_metadata->getMetadata('exif', $test[0]);
-        $this->assertEqual($test[1], $entry ? $entry['value'] : NULL);
+        $this->assertEquals($test[1], $entry ? $entry['value'] : NULL);
       }
     }
 
@@ -143,21 +144,21 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
     $metadata = $file_metadata_from->getMetadata('exif');
     $new_file_metadata = $fmdm->uri('public://test-output.jpeg');
     $new_file_metadata->loadMetadata('exif', $metadata);
-    $this->assertEqual($image_files[0]['count_keys'], $this->countMetadataKeys($new_file_metadata, 'exif'));
+    $this->assertEquals($image_files[0]['count_keys'], $this->countMetadataKeys($new_file_metadata, 'exif'));
     foreach ($image_files[0]['test_keys'] as $test) {
       $entry = $file_metadata->getMetadata('exif', $test[0]);
-      $this->assertEqual($test[1], $new_file_metadata->getMetadata('exif', $test[0])['value']);
+      $this->assertEquals($test[1], $new_file_metadata->getMetadata('exif', $test[0])['value']);
     }
 
     // Test removing metadata.
     $fmdm->release($image_files[0]['uri']);
     $this->assertFalse($fmdm->has($image_files[0]['uri']));
     $file_metadata = $fmdm->uri($image_files[0]['uri']);
-    $this->assertEqual($image_files[0]['count_keys'], $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals($image_files[0]['count_keys'], $this->countMetadataKeys($file_metadata, 'exif'));
     $this->assertTrue($file_metadata->removeMetadata('exif', 'shutterspeedValue'));
     $this->assertTrue($file_metadata->removeMetadata('exif', 'apertureValue'));
     $this->assertFalse($file_metadata->removeMetadata('exif', 'bar'));
-    $this->assertEqual($image_files[0]['count_keys'] - 2, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals($image_files[0]['count_keys'] - 2, $this->countMetadataKeys($file_metadata, 'exif'));
     $this->assertNull($file_metadata->getMetadata('exif', 'shutterspeedValue'));
     $this->assertNull($file_metadata->getMetadata('exif', 'apertureValue'));
     $this->assertNotNull($file_metadata->getMetadata('exif', 'orientation'));
@@ -170,75 +171,84 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
     $fmdm = $this->container->get('file_metadata_manager');
 
     // Copy test file to public://.
-    file_unmanaged_copy(drupal_get_path('module', 'image_effects') . '/tests/images/portrait-painting.jpg', 'public://', FILE_EXISTS_REPLACE);
+    $this->fileSystem->copy(drupal_get_path('module', 'image_effects') . '/tests/images/portrait-painting.jpg', 'public://', FileSystemInterface::EXISTS_REPLACE);
     $file_uri = 'public://portrait-painting.jpg';
     $file_metadata = $fmdm->uri($file_uri);
 
     // Check values via exif_read_data before operations.
     $data = @exif_read_data($file_uri);
-    $this->assertEqual(8, $data['Orientation']);
+    $this->assertEquals(8, $data['Orientation']);
     $this->assertFalse(isset($data['Artist']));
-    $this->assertEqual('Canon', $data['Make']);
-    $this->assertEqual(800, $data['ISOSpeedRatings']);
+    $this->assertEquals('Canon', $data['Make']);
+    $this->assertEquals(800, $data['ISOSpeedRatings']);
 
     // Change the Orientation tag from IFD0.
-    $this->assertEqual(8, $file_metadata->getMetadata('exif', 'orientation')['value']);
-    $this->assertTrue($file_metadata->setMetadata('exif', 'orientation', 4));
-    $this->assertEqual(4, $file_metadata->getMetadata('exif', 'orientation')['value']);
+    $this->assertEquals(8, $file_metadata->getMetadata('exif', 'orientation')['value']);
+    $data['Orientation'] = 4;
+    $this->assertTrue($file_metadata->setMetadata('exif', 'orientation', $data['Orientation']));
+    $this->assertEquals($data['Orientation'], $file_metadata->getMetadata('exif', 'orientation')['value']);
     // Add the Artist tag to IFD0.
-    $this->assertEqual(48, $this->countMetadataKeys($file_metadata, 'exif'));
+    $data['Artist'] = 'shot by foo!';
+    $this->assertEquals(47, $this->countMetadataKeys($file_metadata, 'exif'));
     $this->assertNull($file_metadata->getMetadata('exif', 'artist'));
     $artist_tag = $this->container->get('file_mdm_exif.tag_mapper')->resolveKeyToIfdAndTag('artist');
-    $artist = new PelEntryAscii($artist_tag['tag'], 'shot by foo!');
+    $artist = new PelEntryAscii($artist_tag['tag'], $data['Artist']);
     $file_metadata->setMetadata('exif', 'artist', $artist);
-    $this->assertEqual('shot by foo!', $file_metadata->getMetadata('exif', 'artist')['value']);
-    $this->assertEqual(49, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals($data['Artist'], $file_metadata->getMetadata('exif', 'artist')['value']);
+    $this->assertEquals(48, $this->countMetadataKeys($file_metadata, 'exif'));
     // Setting an unknown tag leads to failure.
     $this->assertFalse($file_metadata->setMetadata('exif', 'bar', 'qux'));
-    // Remove the Make tag from IFD0.
-    $this->assertEqual('Canon', $file_metadata->getMetadata('exif', 'make')['value']);
-    $this->assertTrue($file_metadata->removeMetadata('exif', 'make'));
-    $this->assertNull($file_metadata->getMetadata('exif', 'make'));
-    $this->assertEqual(48, $this->countMetadataKeys($file_metadata, 'exif'));
+    // Remove the ResolutionUnit tag from IFD0.
+    unset($data['ResolutionUnit']);
+    $this->assertEquals(2, $file_metadata->getMetadata('exif', 'ResolutionUnit')['value']);
+    $this->assertTrue($file_metadata->removeMetadata('exif', 'ResolutionUnit'));
+    $this->assertNull($file_metadata->getMetadata('exif', 'ResolutionUnit'));
+    $this->assertEquals(47, $this->countMetadataKeys($file_metadata, 'exif'));
 
     // Add the ImageDescription tag to IFD1.
+    $data['THUMBNAIL']['ImageDescription'] = 'awesome!';
     $this->assertNull($file_metadata->getMetadata('exif', [1, 'imagedescription']));
     $desc_tag = $this->container->get('file_mdm_exif.tag_mapper')->resolveKeyToIfdAndTag([1, 'imagedescription']);
-    $desc = new PelEntryAscii($desc_tag['tag'], 'awesome!');
+    $desc = new PelEntryAscii($desc_tag['tag'], $data['THUMBNAIL']['ImageDescription']);
     $file_metadata->setMetadata('exif', [1, 'imagedescription'], $desc);
-    $this->assertEqual('awesome!', $file_metadata->getMetadata('exif', [1, 'imagedescription'])['value']);
-    $this->assertEqual(49, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals($data['THUMBNAIL']['ImageDescription'], $file_metadata->getMetadata('exif', [1, 'imagedescription'])['value']);
+    $this->assertEquals(48, $this->countMetadataKeys($file_metadata, 'exif'));
     // Remove the Compression tag from IFD1.
-    $this->assertEqual(6, $file_metadata->getMetadata('exif', [1, 'compression'])['value']);
+    unset($data['THUMBNAIL']['Compression']);
+    $this->assertEquals(6, $file_metadata->getMetadata('exif', [1, 'compression'])['value']);
     $this->assertTrue($file_metadata->removeMetadata('exif', [1, 'compression']));
     $this->assertNull($file_metadata->getMetadata('exif', [1, 'compression']));
-    $this->assertEqual(48, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals(47, $this->countMetadataKeys($file_metadata, 'exif'));
 
     // Add the BrightnessValue tag to EXIF.
+    $data['BrightnessValue'] = '12/4';
     $this->assertNull($file_metadata->getMetadata('exif', ['exif', 'brightnessvalue']));
     $brightness_tag = $this->container->get('file_mdm_exif.tag_mapper')->resolveKeyToIfdAndTag(['exif', 'brightnessvalue']);
     $brightness = new PelEntrySRational($brightness_tag['tag'], [12, 4]);
     $file_metadata->setMetadata('exif', ['exif', 'brightnessvalue'], $brightness);
-    $this->assertEqual('12/4', $file_metadata->getMetadata('exif', ['exif', 'brightnessvalue'])['text']);
-    $this->assertEqual(49, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals($data['BrightnessValue'], $file_metadata->getMetadata('exif', ['exif', 'brightnessvalue'])['text']);
+    $this->assertEquals(48, $this->countMetadataKeys($file_metadata, 'exif'));
     // Remove the ISOSpeedRatings tag from EXIF.
-    $this->assertEqual(800, $file_metadata->getMetadata('exif', ['exif', 'isospeedratings'])['value']);
+    unset($data['ISOSpeedRatings']);
+    $this->assertEquals(800, $file_metadata->getMetadata('exif', ['exif', 'isospeedratings'])['value']);
     $this->assertTrue($file_metadata->removeMetadata('exif', ['exif', 'isospeedratings']));
     $this->assertNull($file_metadata->getMetadata('exif', ['exif', 'isospeedratings']));
-    $this->assertEqual(48, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals(47, $this->countMetadataKeys($file_metadata, 'exif'));
 
     // Add the RelatedImageFileFormat tag to INTEROP.
+    $data['RelatedFileFormat'] = 'qux';
     $this->assertNull($file_metadata->getMetadata('exif', ['interop', 'RelatedImageFileFormat']));
     $ff_tag = $this->container->get('file_mdm_exif.tag_mapper')->resolveKeyToIfdAndTag(['interop', 'RelatedImageFileFormat']);
-    $ff = new PelEntryAscii($ff_tag['tag'], 'qux');
+    $ff = new PelEntryAscii($ff_tag['tag'], $data['RelatedFileFormat']);
     $file_metadata->setMetadata('exif', ['interop', 'RelatedImageFileFormat'], $ff);
-    $this->assertEqual('qux', $file_metadata->getMetadata('exif', ['interop', 'RelatedImageFileFormat'])['text']);
-    $this->assertEqual(49, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals($data['RelatedFileFormat'], $file_metadata->getMetadata('exif', ['interop', 'RelatedImageFileFormat'])['text']);
+    $this->assertEquals(48, $this->countMetadataKeys($file_metadata, 'exif'));
     // Remove the InteroperabilityIndex tag from INTEROP.
-    $this->assertEqual('R98', $file_metadata->getMetadata('exif', ['interop', 'InteroperabilityIndex'])['value']);
+    unset($data['InterOperabilityIndex']);
+    $this->assertEquals('R98', $file_metadata->getMetadata('exif', ['interop', 'InteroperabilityIndex'])['value']);
     $this->assertTrue($file_metadata->removeMetadata('exif', ['interop', 'InteroperabilityIndex']));
     $this->assertNull($file_metadata->getMetadata('exif', ['interop', 'InteroperabilityIndex']));
-    $this->assertEqual(48, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals(47, $this->countMetadataKeys($file_metadata, 'exif'));
 
     // Add Longitude/Latitude tags to GPS.
     $this->assertNull($file_metadata->getMetadata('exif', 'GPSLatitudeRef'));
@@ -249,69 +259,79 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
     $at_tag = $this->container->get('file_mdm_exif.tag_mapper')->resolveKeyToIfdAndTag('GPSLatitude');
     $otr_tag = $this->container->get('file_mdm_exif.tag_mapper')->resolveKeyToIfdAndTag('GPSLongitudeRef');
     $ot_tag = $this->container->get('file_mdm_exif.tag_mapper')->resolveKeyToIfdAndTag('GPSLongitude');
-    $atr = new PelEntryAscii($atr_tag['tag'], 'N');
+    $data['GPSLatitudeRef'] = 'N';
+    $atr = new PelEntryAscii($atr_tag['tag'], $data['GPSLatitudeRef']);
+    $data['GPSLatitude'] = ['46/1', '37/1', '59448/10000'];
     $at = new PelEntryRational($at_tag['tag'], [46, 1], [37, 1], [59448, 10000]);
-    $otr = new PelEntryAscii($otr_tag['tag'], 'E');
+    $data['GPSLongitudeRef'] = 'E';
+    $otr = new PelEntryAscii($otr_tag['tag'], $data['GPSLongitudeRef']);
+    $data['GPSLongitude'] = ['12/1', '17/1', '488112/10000'];
     $ot = new PelEntryRational($ot_tag['tag'], [12, 1], [17, 1], [488112, 10000]);
     $file_metadata->setMetadata('exif', 'GPSLatitudeRef', $atr);
     $file_metadata->setMetadata('exif', 'GPSLatitude', $at);
     $file_metadata->setMetadata('exif', 'GPSLongitudeRef', $otr);
     $file_metadata->setMetadata('exif', 'GPSLongitude', $ot);
-    $this->assertEqual('N', $file_metadata->getMetadata('exif', 'GPSLatitudeRef')['text']);
-    $this->assertNotNull($file_metadata->getMetadata('exif', 'GPSLatitude')['text']);
-    $this->assertEqual('E', $file_metadata->getMetadata('exif', 'GPSLongitudeRef')['text']);
-    $this->assertNotNull($file_metadata->getMetadata('exif', 'GPSLongitude')['text']);
-    $this->assertEqual(52, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertEquals($data['GPSLatitudeRef'], $file_metadata->getMetadata('exif', 'GPSLatitudeRef')['text']);
+    $this->assertEquals('46° 37\' 5.9448" (46.62°)', $file_metadata->getMetadata('exif', 'GPSLatitude')['text']);
+    $this->assertEquals($data['GPSLongitudeRef'], $file_metadata->getMetadata('exif', 'GPSLongitudeRef')['text']);
+    $this->assertEquals('12° 17\' 48.8112" (12.30°)', $file_metadata->getMetadata('exif', 'GPSLongitude')['text']);
+    $this->assertEquals(51, $this->countMetadataKeys($file_metadata, 'exif'));
 
     // Save metadata to file.
     $this->assertTrue($file_metadata->saveMetadataToFile('exif'));
 
-    // Check results via exif_read_data.
-    $data = @exif_read_data($file_uri);
-    $this->assertEqual(4, $data['Orientation']);
-    $this->assertEqual('shot by foo!', $data['Artist']);
-    $this->assertFalse(isset($data['Make']));
-    $this->assertEqual('12/4', $data['BrightnessValue']);
-    $this->assertFalse(isset($data['ISOSpeedRatings']));
-    $this->assertEqual('qux', $data['RelatedFileFormat']);
-    $this->assertFalse(isset($data['InterOperabilityIndex']));
-    $this->assertEqual('N', $data['GPSLatitudeRef']);
-    $this->assertEqual(['46/1', '37/1', '59448/10000'], $data['GPSLatitude']);
-    $this->assertEqual('E', $data['GPSLongitudeRef']);
-    $this->assertEqual(['12/1', '17/1', '488112/10000'], $data['GPSLongitude']);
+    // Check results via exif_read_data. A complete check would be through
+    // $this->assertEquals($data, $data_reloaded).
+    $data_reloaded = @exif_read_data($file_uri);
+    $this->assertEquals($data['Orientation'], $data_reloaded['Orientation']);
+    $this->assertEquals($data['Artist'], $data_reloaded['Artist']);
+    $this->assertFalse(isset($data_reloaded['ResolutionUnit']));
+    $this->assertEquals($data['THUMBNAIL']['ImageDescription'], $data_reloaded['THUMBNAIL']['ImageDescription']);
+    $this->assertFalse(isset($data_reloaded['THUMBNAIL']['Compression']));
+    $this->assertEquals($data['BrightnessValue'], $data_reloaded['BrightnessValue']);
+    $this->assertFalse(isset($data_reloaded['ISOSpeedRatings']));
+    $this->assertEquals($data['RelatedFileFormat'], $data_reloaded['RelatedFileFormat']);
+    $this->assertFalse(isset($data_reloaded['InterOperabilityIndex']));
+    $this->assertEquals($data['GPSLatitudeRef'], $data_reloaded['GPSLatitudeRef']);
+    $this->assertEquals($data['GPSLatitude'], $data_reloaded['GPSLatitude']);
+    $this->assertEquals($data['GPSLongitudeRef'], $data_reloaded['GPSLongitudeRef']);
+    $this->assertEquals($data['GPSLongitude'], $data_reloaded['GPSLongitude']);
 
     // Test writing metadata to a file that has no EXIF info.
-    file_unmanaged_copy(drupal_get_path('module', 'simpletest') . '/files/image-2.jpg', 'public://', FILE_EXISTS_REPLACE);
+    $this->fileSystem->copy('core/tests/fixtures/files/image-2.jpg', 'public://', FileSystemInterface::EXISTS_REPLACE);
     $test_2 = $fmdm->uri('public://image-2.jpg');
-    $this->assertEqual(0, $this->countMetadataKeys($test_2, 'exif'));
+    $this->assertEquals(0, $this->countMetadataKeys($test_2, 'exif'));
     // Load EXIF metadata from previous file processed.
     $test_2->loadMetadata('exif', $file_metadata->getMetadata('exif'));
     // Save metadata to file.
     $this->assertTrue($test_2->saveMetadataToFile('exif'));
-    $this->assertEqual(52, $this->countMetadataKeys($test_2, 'exif'));
-    // Check results via exif_read_data.
-    $data = @exif_read_data('public://image-2.jpg');
-    $this->assertEqual(4, $data['Orientation']);
-    $this->assertEqual('shot by foo!', $data['Artist']);
-    $this->assertEqual('12/4', $data['BrightnessValue']);
-    $this->assertEqual('qux', $data['RelatedFileFormat']);
-    $this->assertEqual('N', $data['GPSLatitudeRef']);
-    $this->assertEqual(['46/1', '37/1', '59448/10000'], $data['GPSLatitude']);
-    $this->assertEqual('E', $data['GPSLongitudeRef']);
-    $this->assertEqual(['12/1', '17/1', '488112/10000'], $data['GPSLongitude']);
+    $this->assertEquals(51, $this->countMetadataKeys($test_2, 'exif'));
+
+    // Check results via exif_read_data. A complete check would be through
+    // $this->assertEquals($data, $data_reloaded).
+    $data_reloaded = @exif_read_data('public://image-2.jpg');
+    $this->assertEquals($data['Orientation'], $data_reloaded['Orientation']);
+    $this->assertEquals($data['Artist'], $data_reloaded['Artist']);
+    $this->assertEquals($data['THUMBNAIL']['ImageDescription'], $data_reloaded['THUMBNAIL']['ImageDescription']);
+    $this->assertEquals($data['BrightnessValue'], $data_reloaded['BrightnessValue']);
+    $this->assertEquals($data['RelatedFileFormat'], $data_reloaded['RelatedFileFormat']);
+    $this->assertEquals($data['GPSLatitudeRef'], $data_reloaded['GPSLatitudeRef']);
+    $this->assertEquals($data['GPSLatitude'], $data_reloaded['GPSLatitude']);
+    $this->assertEquals($data['GPSLongitudeRef'], $data_reloaded['GPSLongitudeRef']);
+    $this->assertEquals($data['GPSLongitude'], $data_reloaded['GPSLongitude']);
 
     // Check that after save, file metadata is retrieved from file first time,
     // then from cache in further requests.
     $file_metadata = NULL;
     $this->assertTrue($fmdm->release($file_uri));
     $file_metadata = $fmdm->uri($file_uri);
-    $this->assertEqual(52, $this->countMetadataKeys($file_metadata, 'exif'));
-    $this->assertIdentical(FileMetadataInterface::LOADED_FROM_FILE, $file_metadata->isMetadataLoaded('exif'));
+    $this->assertEquals(51, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertSame(FileMetadataInterface::LOADED_FROM_FILE, $file_metadata->isMetadataLoaded('exif'));
     $file_metadata = NULL;
     $this->assertTrue($fmdm->release($file_uri));
     $file_metadata = $fmdm->uri($file_uri);
-    $this->assertEqual(52, $this->countMetadataKeys($file_metadata, 'exif'));
-    $this->assertIdentical(FileMetadataInterface::LOADED_FROM_CACHE, $file_metadata->isMetadataLoaded('exif'));
+    $this->assertEquals(51, $this->countMetadataKeys($file_metadata, 'exif'));
+    $this->assertSame(FileMetadataInterface::LOADED_FROM_CACHE, $file_metadata->isMetadataLoaded('exif'));
   }
 
   /**
@@ -321,30 +341,44 @@ class FileMetadataExifTest extends FileMetadataManagerTestBase {
     $fmdm = $this->container->get('file_metadata_manager');
 
     // Copy test file to public://.
-    file_unmanaged_copy(drupal_get_path('module', 'file_mdm') . '/tests/files/sample-1.tiff', 'public://', FILE_EXISTS_REPLACE);
+    $this->fileSystem->copy(drupal_get_path('module', 'file_mdm') . '/tests/files/sample-1.tiff', 'public://', FileSystemInterface::EXISTS_REPLACE);
     $file_uri = 'public://sample-1.tiff';
     $file_metadata = $fmdm->uri($file_uri);
 
     // Check values via exif_read_data before operations.
     $data = @exif_read_data($file_uri);
-    $this->assertEqual(1, $data['Orientation']);
-    $this->assertEqual(2, $data['PhotometricInterpretation']);
+    $this->assertEquals(1, $data['Orientation']);
+    $this->assertEquals(2, $data['PhotometricInterpretation']);
+    $this->assertEquals([8, 8, 8, 8], $data['BitsPerSample']);
 
     // Change tags from IFD0.
-    $this->assertEqual(1, $file_metadata->getMetadata('exif', 'orientation')['value']);
-    $this->assertTrue($file_metadata->setMetadata('exif', 'orientation', 4));
-    $this->assertEqual(4, $file_metadata->getMetadata('exif', 'orientation')['value']);
-    $this->assertEqual(2, $file_metadata->getMetadata('exif', 'PhotometricInterpretation')['value']);
-    $this->assertTrue($file_metadata->setMetadata('exif', 'PhotometricInterpretation', 4));
-    $this->assertEqual(4, $file_metadata->getMetadata('exif', 'PhotometricInterpretation')['value']);
+    $this->assertEquals(15, $this->countMetadataKeys($file_metadata, 'exif'));
+
+    $this->assertEquals(1, $file_metadata->getMetadata('exif', 'orientation')['value']);
+    $data['Orientation'] = 4;
+    $this->assertTrue($file_metadata->setMetadata('exif', 'orientation', $data['Orientation']));
+    $this->assertEquals($data['Orientation'], $file_metadata->getMetadata('exif', 'orientation')['value']);
+
+    $this->assertEquals(2, $file_metadata->getMetadata('exif', 'PhotometricInterpretation')['value']);
+    $data['PhotometricInterpretation'] = 4;
+    $this->assertTrue($file_metadata->setMetadata('exif', 'PhotometricInterpretation', $data['PhotometricInterpretation']));
+    $this->assertEquals($data['PhotometricInterpretation'], $file_metadata->getMetadata('exif', 'PhotometricInterpretation')['value']);
+
+    $this->assertEquals([8, 8, 8, 8], $file_metadata->getMetadata('exif', 'BitsPerSample')['value']);
+    $data['BitsPerSample'] = [7, 6, 5, 4];
+    $bps_tag = $this->container->get('file_mdm_exif.tag_mapper')->resolveKeyToIfdAndTag('BitsPerSample');
+    $bps = new PelEntryShort($bps_tag['tag'], $data['BitsPerSample']);
+    $this->assertTrue($file_metadata->setMetadata('exif', 'BitsPerSample', $bps));
+    $this->assertEquals($data['BitsPerSample'], $file_metadata->getMetadata('exif', 'BitsPerSample')['value']);
 
     // Save metadata to file.
     $this->assertTrue($file_metadata->saveMetadataToFile('exif'));
 
     // Check results via exif_read_data.
-    $data = @exif_read_data($file_uri);
-    $this->assertEqual(4, $data['Orientation']);
-    $this->assertEqual(4, $data['PhotometricInterpretation']);
+    $data_reloaded = @exif_read_data($file_uri);
+    $this->assertEquals(4, $data_reloaded['Orientation']);
+    $this->assertEquals(4, $data_reloaded['PhotometricInterpretation']);
+    $this->assertEquals([7, 6, 5, 4], $data_reloaded['BitsPerSample']);
   }
 
 }
