@@ -121,7 +121,7 @@ class FileMetadata implements FileMetadataInterface {
       $this->fileSystem->unlink($temp_uri);
       $temp_uri .= '.' . pathinfo($this->getUri(), PATHINFO_EXTENSION);
     }
-    if ($temp_path = file_unmanaged_copy($this->getUri(), $this->fileSystem->realpath($temp_uri), FILE_EXISTS_REPLACE)) {
+    if ($temp_path = $this->fileSystem->copy($this->getUri(), $this->fileSystem->realpath($temp_uri), FileSystemInterface::EXISTS_REPLACE)) {
       $this->setLocalTempPath($temp_path);
     }
     return (bool) $temp_path;
@@ -134,7 +134,7 @@ class FileMetadata implements FileMetadataInterface {
     if (($temp_path = $this->getLocalTempPath()) === NULL) {
       return FALSE;
     }
-    return (bool) file_unmanaged_copy($temp_path, $this->getUri(), FILE_EXISTS_REPLACE);
+    return (bool) $this->fileSystem->copy($temp_path, $this->getUri(), FileSystemInterface::EXISTS_REPLACE);
   }
 
   /**
