@@ -5,6 +5,7 @@ namespace Drupal\file_mdm;
 use Drupal\Core\Cache\CacheBackendInterface;
 use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Core\File\FileSystemInterface;
+use Drupal\Core\StreamWrapper\StreamWrapperManagerInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\file_mdm\Plugin\FileMetadataPluginManager;
 use Psr\Log\LoggerInterface;
@@ -52,6 +53,13 @@ class FileMetadataManager implements FileMetadataManagerInterface {
   protected $cache;
 
   /**
+   * The stream wrapper manager service.
+   *
+   * @var \Drupal\Core\StreamWrapper\StreamWrapperManagerInterface
+   */
+  protected $streamWrapperManager;
+
+  /**
    * The array of FileMetadata objects currently in use.
    *
    * @var \Drupal\file_mdm\FileMetadataInterface[]
@@ -71,13 +79,16 @@ class FileMetadataManager implements FileMetadataManagerInterface {
    *   The file system service.
    * @param \Drupal\Core\Cache\CacheBackendInterface $cache_service
    *   The cache service.
+   * @param \Drupal\Core\StreamWrapper\StreamWrapperManagerInterface $stream_wrapper_manager
+   *   The stream wrapper manager service.
    */
-  public function __construct(FileMetadataPluginManager $plugin_manager, LoggerInterface $logger, ConfigFactoryInterface $config_factory, FileSystemInterface $file_system, CacheBackendInterface $cache_service) {
+  public function __construct(FileMetadataPluginManager $plugin_manager, LoggerInterface $logger, ConfigFactoryInterface $config_factory, FileSystemInterface $file_system, CacheBackendInterface $cache_service, StreamWrapperManagerInterface $stream_wrapper_manager) {
     $this->pluginManager = $plugin_manager;
     $this->logger = $logger;
     $this->configFactory = $config_factory;
     $this->fileSystem = $file_system;
     $this->cache = $cache_service;
+    $this->streamWrapperManager = $stream_wrapper_manager;
   }
 
   /**
@@ -94,7 +105,7 @@ class FileMetadataManager implements FileMetadataManagerInterface {
     // @see http://stackoverflow.com/questions/12494515/remove-unnecessary-slashes-from-path
     $uri = preg_replace('/([^:])(\/{2,})/', '$1/', $uri);
     // If URI is invalid and no local file path exists, return NULL.
-    if (!file_valid_uri($uri) && !$this->fileSystem->realpath($uri)) {
+    if (!$this->streamWrapperManager->isValidUri($uri) && !$this->fileSystem->realpath($uri)) {
       return NULL;
     }
     // Return a hash of the URI.
@@ -106,7 +117,7 @@ class FileMetadataManager implements FileMetadataManagerInterface {
    */
   public function has($uri) {
     $hash = $this->calculateHash($uri);
-    return $hash ? isset($this->files[$hash]) : NULL;
+    return $hash ? isset($this->files[$hash]) : FALSE;
   }
 
   /**
