@@ -4,11 +4,14 @@ namespace Drupal\honeypot_test\Controller;
 
 use Drupal\Core\Form\FormState;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
 
 /**
  * Controller for honeypot_test routes.
  */
 class HoneypotTestController {
+
+  use StringTranslationTrait;
 
   /**
    * Page that triggers a programmatic form submission.
@@ -20,7 +23,7 @@ class HoneypotTestController {
     $values = [
       'name' => 'robo-user',
       'mail' => 'robouser@example.com',
-      'op' => t('Submit'),
+      'op' => $this->t('Submit'),
     ];
     $form_state->setValues($values);
     \Drupal::formBuilder()->submitForm('\Drupal\user\Form\UserPasswordForm', $form_state);
