@@ -3,6 +3,7 @@
 namespace Drupal\Tests\honeypot\Functional;
 
 use Drupal\Tests\BrowserTestBase;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
 
 /**
  * Test Honeypot spam protection admin form functionality.
@@ -10,6 +11,8 @@ use Drupal\Tests\BrowserTestBase;
  * @group honeypot
  */
 class HoneypotAdminFormTest extends BrowserTestBase {
+
+  use StringTranslationTrait;
 
   /**
    * Admin user.
@@ -55,14 +58,14 @@ class HoneypotAdminFormTest extends BrowserTestBase {
 
     // Set up form and submit it.
     $edit['element_name'] = "test";
-    $this->drupalPostForm('admin/config/content/honeypot', $edit, t('Save configuration'));
+    $this->drupalPostForm('admin/config/content/honeypot', $edit, $this->t('Save configuration'));
 
     // Form should have been submitted successfully.
     $this->assertSession()->pageTextContains('The configuration options have been saved.');
 
     // Set up form and submit it.
     $edit['element_name'] = "test-1";
-    $this->drupalPostForm('admin/config/content/honeypot', $edit, t('Save configuration'));
+    $this->drupalPostForm('admin/config/content/honeypot', $edit, $this->t('Save configuration'));
 
     // Form should have been submitted successfully.
     $this->assertSession()->pageTextContains('The configuration options have been saved.');
@@ -77,7 +80,7 @@ class HoneypotAdminFormTest extends BrowserTestBase {
 
     // Set up form and submit it.
     $edit['element_name'] = "1test";
-    $this->drupalPostForm('admin/config/content/honeypot', $edit, t('Save configuration'));
+    $this->drupalPostForm('admin/config/content/honeypot', $edit, $this->t('Save configuration'));
 
     // Form submission should fail.
     $this->assertSession()->pageTextContains('The element name must start with a letter.');
@@ -92,14 +95,14 @@ class HoneypotAdminFormTest extends BrowserTestBase {
 
     // Set up form and submit it.
     $edit['element_name'] = "special-character-&";
-    $this->drupalPostForm('admin/config/content/honeypot', $edit, t('Save configuration'));
+    $this->drupalPostForm('admin/config/content/honeypot', $edit, $this->t('Save configuration'));
 
     // Form submission should fail.
     $this->assertSession()->pageTextContains('The element name cannot contain spaces or other special characters.');
 
     // Set up form and submit it.
     $edit['element_name'] = "space in name";
-    $this->drupalPostForm('admin/config/content/honeypot', $edit, t('Save configuration'));
+    $this->drupalPostForm('admin/config/content/honeypot', $edit, $this->t('Save configuration'));
 
     // Form submission should fail.
     $this->assertSession()->pageTextContains('The element name cannot contain spaces or other special characters.');
