@@ -5,6 +5,7 @@ namespace Drupal\Tests\honeypot\Functional;
 use Drupal\comment\Tests\CommentTestTrait;
 use Drupal\comment\Plugin\Field\FieldType\CommentItemInterface;
 use Drupal\contact\Entity\ContactForm;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\Tests\BrowserTestBase;
 use Drupal\user\UserInterface;
 
@@ -16,6 +17,7 @@ use Drupal\user\UserInterface;
 class HoneypotFormTest extends BrowserTestBase {
 
   use CommentTestTrait;
+  use StringTranslationTrait;
 
   /**
    * Admin user.
@@ -124,7 +126,7 @@ class HoneypotFormTest extends BrowserTestBase {
     // Set up form and submit it.
     $edit['name'] = $this->randomMachineName();
     $edit['mail'] = $edit['name'] . '@example.com';
-    $this->drupalPostForm('user/register', $edit, t('Create new account'));
+    $this->drupalPostForm('user/register', $edit, $this->t('Create new account'));
 
     // Form should have been submitted successfully.
     $this->assertSession()->pageTextContains('A welcome message with further instructions has been sent to your email address.');
@@ -138,7 +140,7 @@ class HoneypotFormTest extends BrowserTestBase {
     $edit['name'] = $this->randomMachineName();
     $edit['mail'] = $edit['name'] . '@example.com';
     $edit['url'] = 'http://www.example.com/';
-    $this->drupalPostForm('user/register', $edit, t('Create new account'));
+    $this->drupalPostForm('user/register', $edit, $this->t('Create new account'));
 
     // Form should have error message.
     $this->assertSession()->pageTextContains('There was a problem with your form submission. Please refresh the page and try again.');
@@ -155,8 +157,8 @@ class HoneypotFormTest extends BrowserTestBase {
     $edit['mail'] = $edit['name'] . '@example.com';
     $this->drupalGet('user/register');
     sleep(2);
-    $this->drupalPostForm(NULL, $edit, t('Create new account'));
-    $this->assertNoText(t('There was a problem with your form submission.'));
+    $this->drupalPostForm(NULL, $edit, $this->t('Create new account'));
+    $this->assertNoText($this->t('There was a problem with your form submission.'));
 
     // Set the time limit a bit higher so we can trigger honeypot.
     \Drupal::configFactory()->getEditable('honeypot.settings')->set('time_limit', 5)->save();
@@ -164,7 +166,7 @@ class HoneypotFormTest extends BrowserTestBase {
     // Set up form and submit it.
     $edit['name'] = $this->randomMachineName();
     $edit['mail'] = $edit['name'] . '@example.com';
-    $this->drupalPostForm('user/register', $edit, t('Create new account'));
+    $this->drupalPostForm('user/register', $edit, $this->t('Create new account'));
 
     // Form should have error message.
     $this->assertSession()->pageTextContains('There was a problem with your form submission. Please wait 6 seconds and try again.');
@@ -181,8 +183,8 @@ class HoneypotFormTest extends BrowserTestBase {
     // Any value that is not strictly empty should trigger Honeypot.
     foreach (['0', ' '] as $value) {
       $edit['url'] = $value;
-      $this->drupalPostForm('user/register', $edit, t('Create new account'));
-      $this->assertText(t('There was a problem with your form submission. Please refresh the page and try again.'), "Honeypot protection is triggered when the honeypot field contains '{$value}'.");
+      $this->drupalPostForm('user/register', $edit, $this->t('Create new account'));
+      $this->assertText($this->t('There was a problem with your form submission. Please refresh the page and try again.'), "Honeypot protection is triggered when the honeypot field contains '{$value}'.");
     }
   }
 
@@ -200,7 +202,7 @@ class HoneypotFormTest extends BrowserTestBase {
 
     // Set up form and submit it.
     $edit["comment_body[0][value]"] = $comment;
-    $this->drupalPostForm('comment/reply/node/' . $this->node->id() . '/comment', $edit, t('Save'));
+    $this->drupalPostForm('comment/reply/node/' . $this->node->id() . '/comment', $edit, $this->t('Save'));
     $this->assertSession()->pageTextContains('Your comment has been queued for review');
   }
 
@@ -216,7 +218,7 @@ class HoneypotFormTest extends BrowserTestBase {
     // Set up form and submit it.
     $edit["comment_body[0][value]"] = $comment;
     $edit['url'] = 'http://www.example.com/';
-    $this->drupalPostForm('comment/reply/node/' . $this->node->id() . '/comment', $edit, t('Save'));
+    $this->drupalPostForm('comment/reply/node/' . $this->node->id() . '/comment', $edit, $this->t('Save'));
     $this->assertSession()->pageTextContains('There was a problem with your form submission. Please refresh the page and try again.');
   }
 
@@ -244,7 +246,7 @@ class HoneypotFormTest extends BrowserTestBase {
 
     // Set up the form and submit it.
     $edit["title[0][value]"] = 'Test Page';
-    $this->drupalPostForm('node/add/article', $edit, t('Save'));
+    $this->drupalPostForm('node/add/article', $edit, $this->t('Save'));
     $this->assertSession()->pageTextContains('There was a problem with your form submission.');
   }
 
@@ -257,7 +259,7 @@ class HoneypotFormTest extends BrowserTestBase {
 
     // Post a node form using the 'Preview' button and make sure it's allowed.
     $edit["title[0][value]"] = 'Test Page';
-    $this->drupalPostForm('node/add/article', $edit, t('Preview'));
+    $this->drupalPostForm('node/add/article', $edit, $this->t('Preview'));
     $this->assertSession()->pageTextNotContains('There was a problem with your form submission.');
   }
 
@@ -288,7 +290,7 @@ class HoneypotFormTest extends BrowserTestBase {
     // Submit the admin form so we can verify the right forms are displayed.
     $this->drupalPostForm('admin/config/content/honeypot', [
       'form_settings[contact_message_feedback_form]' => TRUE,
-    ], t('Save configuration'));
+    ], $this->t('Save configuration'));
 
     $this->drupalLogin($this->webUser);
     $this->drupalGet('contact/feedback');
