@@ -14,7 +14,7 @@ class ExecManagerTest extends KernelTestBase {
   /**
    * {@inheritdoc}
    */
-  public static $modules = ['imagemagick'];
+  public static $modules = ['imagemagick', 'file_mdm', 'sophron'];
 
   /**
    * Test missing command on ExecManager.
